Second review: исправлены критические ошибки - Эффект остановлен, если форма закрыта. - Весь главный контент страницы помещен в тег main. - Компонент App вынесен в отдельный файл. - clsx можно не используется, если только один селектор передается.
